Python/requirements.txt
app==0.0.1
bcrypt==3.1.7
bson==0.5.10
cffi==1.14.0
click==7.1.2
dlib==19.20.0
dnspython==1.16.0
email-validator==1.1.1
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.58.0
fastapi-mail==0.2.5
Flask==1.1.2
gevent==20.6.2
greenlet==0.4.16
h11==0.9.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
motor==2.1.0
numpy==1.19.0
opencv-python==4.2.0.34
passlib==1.7.2
Pillow==7.1.2
pycparser==2.20
pydantic==1.5.1
PyJWT==1.7.1
pymongo==3.10.1
python-dateutil==2.8.1
python-multipart==0.0.5
six==1.15.0
starlette==0.13.4
uvicorn==0.11.5
websockets==8.1
Werkzeug==1.0.1
zope.event==4.4
zope.interface==5.1.0
